feature:Added Permissions to individual routes

+++ routes/RoleAndPermission/Role.php
@@ -5,7 +5,7 @@ use App\Http\Controllers\RoleController;
 use Illuminate\Support\Facades\Route;
 
 // Create a new role
-Route::post('/roles', [RoleController::class, 'createRole'])
+Route::middleware(['permission:appAdmin.role.create'])->post('/roles', [RoleController::class, 'createRole'])
 ->name('roles.store');
 
 // Get all roles
@@ -13,18 +13,18 @@ Route::get('/roles', [RoleController::class, 'getRoles'])
 ->name('roles.index');
 
 // Delete a specific role
-Route::delete('/roles/{roleId}', [RoleController::class, 'deleteRole']) // Assuming 'updateRole' in delete route was a typo
+Route::middleware(['permission:appAdmin.role.delete'])->delete('/roles/{roleId}', [RoleController::class, 'deleteRole']) // Assuming 'updateRole' in delete route was a typo
 ->name('roles.destroy');
 
 // Update a specific role
-Route::put('/roles/{roleId}', [RoleController::class, 'updateRole'])
+Route::middleware(['permission:appAdmin.role.update'])->put('/roles/{roleId}', [RoleController::class, 'updateRole'])
 ->name('roles.update');
 
 // Assign a role to a school administrator
-Route::middleware([ IdentifyTenant::class])->post('/school-admins/{schoolAdminId}/roles', [RoleController::class, 'assignRoleSchoolAdmin'])
+Route::middleware([ IdentifyTenant::class, 'permission:schoolAdmin.role.assign'])->post('/school-admins/{schoolAdminId}/roles', [RoleController::class, 'assignRoleSchoolAdmin'])
 ->name('school-admins.roles.store'); // Using 'store' as it creates an association
 
 // Remove a role from a school administrator
-Route::post('/school-admins/{schoolAdminId}/roles', [RoleController::class, 'removeRoleSchoolAdmin'])
+Route::middleware(['permission:schoolAdmin.role.remove'])->post('/school-admins/{schoolAdminId}/roles', [RoleController::class, 'removeRoleSchoolAdmin'])
 ->name('school-admins.roles.destroy');
